code cleanup and adds localization
