+++ src/components/primitive/menu-component/menu.component.props.ts
@@ -1,7 +1,4 @@
-import { ButtonComponentProps } from '..';
-import { TextComponentProps } from '../text/text.component.props';
-
-export interface GenericModalComponentProps {
+export interface MenuComponentProps {
   /**
    * to determine if modal should show
    */
@@ -11,26 +8,6 @@ export interface GenericModalComponentProps {
    * method to close the modal
    */
   onClose?: () => void;
-
-  /**
-   * title of the modal
-   */
-  title?: TextComponentProps;
-
-  /**
-   * button for submit
-   */
-  onSubmit?: ButtonComponentProps;
-
-  /**
-   * button for cancel
-   */
-  onCancel?: ButtonComponentProps;
-
-  /**
-   * hide the buttons and close button
-   */
-  hide?: boolean;
 
   /**
    * should modal close if background is clicked
