+++ src/components/primitive/top-bar/top-bar.component.tsx
@@ -1,15 +1,13 @@
+import { useState } from 'react';
+import { Link } from 'react-router-dom';
+import { RoutePath } from 'navigation/route-path';
 import { Image, Avatar } from 'components/primitive';
 import { TopBarContainer, LogoWrapper, BoxGroup, NotificationIcon } from './top-bar.component.styles';
-import { RoutePath } from 'navigation/route-path';
-import { Link } from 'react-router-dom';
 import { Menu } from '../menu-component/menu.component';
-import { useState } from 'react';
-import { GenericModal } from '../generic-modal/generic-modal.component';
 
 export const TopBar = () => {
   const [showMenu, setShowMenu] = useState(false);
 
-  console.log('menu', showMenu);
   return (
     <>
       <TopBarContainer align={'center'} justify={'center'} hCenter vCenter row>
@@ -20,8 +18,7 @@ export const TopBar = () => {
         </LogoWrapper>
         <BoxGroup row>
           <NotificationIcon name='notification' size='small' />
-
-          <Avatar name='user' size='small' onClick={() => setShowMenu(!showMenu)} />
+          <Avatar name='user' size='small' onClick={() => setShowMenu(!showMenu)} style={{ cursor: 'pointer' }} />
         </BoxGroup>
       </TopBarContainer>
       {showMenu && <Menu show />}
