+++ src/themes/images.ts
@@ -13,7 +13,7 @@ export const images = {
   claim: require('../assets/icons/claims.svg').default,
   close: require('../assets/icons/close.svg').default,
   copy: require('../assets/icons/copy.svg').default,
-  darkMode: require('../assets/icons/darkMode.svg').default,
+  darkMode: require('../assets/icons/dark-mode.svg').default,
   dashboard: require('../assets/icons/dashboard.svg').default,
   dayMode: require('../assets/icons/dayMode.svg').default,
   discord: require('../assets/icons/discord.svg').default,
@@ -57,6 +57,8 @@ export const images = {
   lockLoader: require('../assets/icons/lockLoader.gif').default,
   safientWalletLogo: require('../assets/logo/logo.svg').default,
   settings: require('../assets/icons/settings.svg').default,
+  settingsDark: require('../assets/icons/settings-dark.svg').default,
+  logout: require('../assets/icons/logout.svg').default,
   receive: require('../assets/icons/recieve.svg').default,
   send: require('../assets/icons/send.svg').default,
   trade: require('../assets/icons/trade.svg').default,
